requirements.txt
# ============ FASTAPI 101 PROJECT DEPENDENCIES ============
#
# This file lists all Python packages required for the FastAPI social media API.
# 
# To install: pip install -r requirements.txt
# To freeze current environment: pip freeze > requirements.txt
#
# Package categorization:
#   [WEB FRAMEWORK] FastAPI, Uvicorn, Python-multipart
#   [DATABASE] SQLAlchemy, Alembic, psycopg2
#   [AUTHENTICATION] python-jose, passlib, bcrypt, argon2-cffi
#   [VALIDATION] Pydantic
#   [UTILITIES] python-dotenv, etc.

# ============ WEB FRAMEWORK & SERVER ============
fastapi==0.135.1                # Modern async web framework for building REST APIs
# FastAPI features: Automatic API docs (/docs), Pydantic validation, dependency injection

uvicorn==0.31.1                 # ASGI server (runs FastAPI app)
# Command: uvicorn app.main:app --reload (runs with auto-reload on code changes)

fastapi-cli==0.0.24             # CLI tools for FastAPI
fastapi-cloud-cli==0.15.0       # Cloud deployment CLI

python-multipart==0.0.22        # Parse multipart form data (for file uploads, form submissions)

# ============ DATABASE & ORM ============
SQLAlchemy==2.0.25              # SQL toolkit and ORM (Object-Relational Mapping)
# Maps Python classes to database tables, handles queries and relationships

psycopg2==2.9.11                # PostgreSQL database adapter (enables SQLAlchemy to talk to PostgreSQL)

alembic==1.18.4                 # Database migration tool for SQLAlchemy
# Tracks schema changes, enables version control of database structure
# Usage: alembic revision --autogenerate -m "description" && alembic upgrade head

green==3.3.2                    # Connection pooling for better performance

# ============ PASSWORD & SECURITY ============
bcrypt==5.0.0                   # Password hashing (older, used by passlib)

passlib==1.7.4                  # Password hashing/verification framework
# Supports multiple algorithms (bcrypt, scrypt, argon2)

argon2-cffi==25.1.0             # Argon2 password hashing (recommended, more secure than bcrypt)
# Memory-hard algorithm resistant to GPU attacks

python-jose==3.5.0              # JWT token library (JSON Web Tokens)
#Used for creating and verifying access tokens

cryptography==46.0.5            # Cryptographic algorithms (used by python-jose)

ecdsa==0.19.1                   # Elliptic Curve Digital Signature (alternative to RSA)

rsa==4.9.1                      # RSA cryptography (for asymmetric encryption)

# ============ DATA VALIDATION & SERIALIZATION ============
pydantic==2.12.5                # Data validation and serialization using Python type hints
# Validates API request/response data, auto-generates API docs

pydantic-settings==2.13.1       # Configuration management using Pydantic
# Load settings from .env file with type validation

pydantic-extra-types==2.11.0    # Additional Pydantic validators (custom validation rules)

email-validator==2.3.0          # Email address validation with Pydantic support
# Used by EmailStr for strict email format checking

annotated-types==0.7.0          # Annotations for type hints (metadata for type checking)
annotated-doc==0.0.4            # Documentation for annotated types

# ============ ENVIRONMENT & CONFIGURATION ============
python-dotenv==1.2.2            # Load environment variables from .env file
# Usage: from dotenv import load_dotenv; load_dotenv()

# ============ HTTP & NETWORKING ============
httpx==0.28.1                   # HTTP client (async/sync) for testing and HTTP calls
httpcore==1.0.9                 # Low-level HTTP library
h11==0.16.0                     # HTTP/1.1 state machine
httptools==0.7.1                # HTTP parsing library (faster than pure Python)

anyio==4.12.1                   # Async I/O abstraction (works with asyncio, trio, etc.)

certifi==2026.2.25              # Mozilla's CA bundle (SSL certificate verification)

idna==3.11                      # Internationalized domain names

# ============ DNS & NETWORKING ============
dnspython==2.8.0                # DNS toolkit (supports async DNS queries)

# ============ UTILITIES ============
click==8.3.1                    # Command-line interface creation (CLI utility)

colorama==0.4.6                 # ANSI color codes for terminal output
Pygments==2.19.2                # Syntax highlighting library (for code display)

Jinja2==3.1.6                   # Template engine (used in Alembic)
MarkupSafe==3.0.3               # Safe string handling for templates
Mako==1.3.10                    # Template library (used by Alembic for migrations)

markdown-it-py==4.0.0           # Markdown parser
mdurl==0.1.2                    # URL parsing for markdown
rich==14.3.3                    # Rich terminal output formatting
rich-toolkit==0.19.7            # Toolkit for Rich library

itsdangerous==2.2.2             # Secure signing of data (used for tokens, cookies)

pyasn1==0.6.3                   # ASN.1 data structures

pycparser==3.0                  # C parser (used by cryptography)
cffi==2.0.0                     # C Foreign Function Interface

pytz==2026.1.post1              # Timezone database
exceptiongroup==1.3.1           # Exception grouping (backport for Python <3.11)

# ============ OPTIONAL / LESS COMMONLY USED ============
fastar==0.8.0                   # Fast async utilities
rignore==0.7.6                  # Ignore file parser
PyYAML==6.0.3                   # YAML format support
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
tomli==2.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
